+++ src/Model/Invoices/IssueData.php
@@ -22,6 +22,7 @@ class IssueData
     public string $projectId;
     public ?int $planHours;
     public ?int $hourRemaining;
+    public ?\DateTime $dueDate = null;
 
     public function __construct()
     {