+++ src/Service/DataSynchronizationService.php
@@ -235,6 +235,7 @@ class DataSynchronizationService
                 $issue->setStatus($issueDatum->status);
                 $issue->setPlanHours($issueDatum->planHours);
                 $issue->setHoursRemaining($issueDatum->hourRemaining);
+                $issue->setDueDate($issueDatum->dueDate);
 
                 // Leantime (as of now) supports only a single version (milestone) per issue.
                 if (LeantimeApiService::class === $dataProvider?->getClass()) {